% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{with_http_auth}
\alias{with_http_auth}
\title{Add HTTP Basic authentication to a shinyApp object}
\usage{
with_http_auth(app, cred)
}
\arguments{
\item{app}{shinyApp object}

\item{cred}{named list of username/password pairs}
}
\description{
Example usage: with_http_auth(shinyApp(ui=ui, server=server), list(user="pw"))
}
